+++ BabyGotBackup/src/edu/vtc/cis4150/ScheduledSession.java
@@ -156,7 +156,8 @@ public class ScheduledSession implements Session{
 				if(_isCompressed)
 					ext = ".zip";
 				File deleteMe = new File(_backupLocation + File.separator + f.getName() + ext);
-				Files.delete(deleteMe.toPath());
+				if (deleteMe.exists())
+					Files.delete(deleteMe.toPath());
 							
 				for (Map.Entry<File, File> fileEnt: _backupToFile.entrySet()) { //foreach for a map. messy, but works
 					if (fileEnt.getKey().getName().equals(deleteMe.getName()))
@@ -518,7 +519,7 @@ public class ScheduledSession implements Session{
 			newSession.setContinueVal(false);
 			
 			_versionNum++;
-			_scheduledBackupTime.roll(Calendar.HOUR, _interval);
+			_scheduledBackupTime.add(Calendar.HOUR, _interval);
 			_lastModifiedDate = Calendar.getInstance();
 			repOK();
